+++ torch/nn/quantized/modules/conv.py
@@ -16,8 +16,6 @@ from torch.nn.quantized.modules.utils import _pair_from_first
 from torch.nn.quantized.modules.utils import _quantize_weight
 from torch.nn.utils import fuse_conv_bn_weights
 
-# from torch.quantization.fx.utils import assert_and_get_unique_device
-
 _SUPPORTED_PADDING = {
     'zeros',
     'reflect'
@@ -207,8 +205,8 @@ class _ConvNd(nn.Module):
                     mod.stride, mod.padding, mod.dilation, mod.groups,
                     mod.bias is not None, mod.padding_mode)
         qconv.set_weight_bias(qweight, mod.bias)
-        qconv.scale = act_scale.to(torch.float)
-        qconv.zero_point = act_zp.to(torch.long)
+        qconv.scale = float(act_scale)
+        qconv.zero_point = int(act_zp)
         return qconv
 
     @staticmethod
